Revert the `mut` keyword
Having it requires handling a lot more semantics on the language level
that I don't want to manage. One of the things I'd like is not exposing
anything regarding mutability or storage to the language. The primary
reason for having mutability had to do with reusing `alloca`s to avoid
excess memcpys when calling functions that returned aggregate types, but
LLVM has an optimization that takes care of this. That is, if all uses
of an alloca occur before the first use of another alloca that's the
same size and type, then the pass will remove the second alloca and
reuse the first one.

Also add --help to the arg parser.